refactor(cli): extract app interaction policy
