+++ src/test-kernel/cli/TuiStateAndFocus.vitest.mts
@@ -26,19 +26,6 @@ import {
   staticTranscriptRowBudget,
 } from '@cli/chat/tui/appLayout';
 import {
-  appEscapeInterruptActive,
-  appFocusShortcutsActive,
-  approvalForegroundMaxRows,
-  approvalVisibleForActiveStream,
-  childControlForegroundMaxRows,
-  digitFromMetaShortcut,
-  foregroundEscapeAction,
-  foregroundSurfaceKind,
-  shouldDeferEscapeInterruptForMetaChord,
-  triggerEscapeInterrupt,
-} from '@cli/chat/tui/App';
-import type { PendingApproval } from '@cli/chat/tui/state/approvalQueue';
-import {
   nextFocusBack,
   nextFocusForward,
 } from '@cli/chat/tui/state/focusCycle';
@@ -439,125 +426,6 @@ describe('CLI TUI row allocation', () => {
     expect(layout.foregroundRows).toBe(18);
   });
 
-  it('uses a smaller row cap for empty child-control pickers', () => {
-    expect(childControlForegroundMaxRows({ hasItems: false })).toBe(6);
-    expect(childControlForegroundMaxRows({ hasItems: true })).toBe(12);
-  });
-
-  it('does not cap natural-height approval prompts that lack row budgeting', () => {
-    const decide = () => {};
-
-    expect(
-      approvalForegroundMaxRows({
-        payload: {
-          kind: 'plan',
-          payload: {
-            approvalId: 'plan-1',
-            streamId: root,
-            goalEnabled: false,
-            plan: { objective: 'Coordinate a proof.' },
-          },
-        },
-        decide,
-      }),
-    ).toBeUndefined();
-    expect(
-      approvalForegroundMaxRows({
-        payload: {
-          kind: 'retry',
-          payload: { streamId: root, operation: 'Model invocation' },
-        },
-        decide,
-      }),
-    ).toBeUndefined();
-    expect(
-      approvalForegroundMaxRows({
-        payload: {
-          kind: 'userQuestion',
-          payload: {
-            requestId: 'question-1',
-            allowBypass: false,
-            streamId: root,
-            context: 'Pick a proof style.',
-            questions: [
-              {
-                question: 'Which style?',
-                options: [{ label: 'direct' }, { label: 'detailed' }],
-              },
-            ],
-          },
-        },
-        decide,
-      }),
-    ).toBeUndefined();
-    expect(
-      approvalForegroundMaxRows({
-        payload: {
-          kind: 'bash',
-          payload: {
-            requestId: 'bash-1',
-            allowBypass: true,
-            streamId: root,
-            command: 'echo ok',
-          },
-        },
-        decide,
-      }),
-    ).toBe(18);
-    expect(
-      approvalForegroundMaxRows({
-        payload: {
-          kind: 'toolEdit',
-          payload: {
-            path: 'draft.tex',
-            originalContent: 'old',
-            proposedContent: 'new',
-            sourceTool: 'test',
-            streamId: root,
-          },
-        },
-        decide,
-      }),
-    ).toBe(18);
-    expect(
-      approvalForegroundMaxRows({
-        payload: {
-          kind: 'proposal',
-          payload: {
-            proposalId: 'proposal-1',
-            streamId: root,
-            agentCategory: AgentCategory.ToolUse,
-            agent: 'review',
-            model: 'harness-model',
-            instruction: 'Review the proof.',
-            memories: [],
-            workingDirectory: '/tmp',
-          },
-        },
-        decide,
-      }),
-    ).toBe(18);
-    expect(
-      approvalForegroundMaxRows({
-        payload: {
-          kind: 'externalInquiry',
-          payload: {
-            requestId: 'inquiry-1',
-            mode: 'new' as const,
-            threadId: 'ei_000000000001',
-            allowBypass: false,
-            streamId: root,
-            question: 'Can you check this claim?',
-            sessionLinks: null,
-            draft: null,
-            transcript: null,
-          },
-        },
-        decide,
-      }),
-    ).toBe(18);
-  });
-
   it('uses the whole middle region for the transcript without foreground UI', () => {
     const layout = allocateMiddleRows({
       foregroundOpen: false,
@@ -793,365 +661,6 @@ describe('CLI TUI row allocation', () => {
         ],
       }),
     ).toBe(false);
-  });
-
-  it('lets input overlays own focus shortcuts', () => {
-    expect(
-      appFocusShortcutsActive({
-        foregroundOpen: false,
-        reverseSearchOpen: false,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(true);
-    expect(
-      appFocusShortcutsActive({
-        foregroundOpen: false,
-        reverseSearchOpen: true,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(false);
-    expect(
-      appFocusShortcutsActive({
-        foregroundOpen: false,
-        reverseSearchOpen: false,
-        slashPaletteOpen: true,
-      }),
-    ).toBe(false);
-    expect(
-      appFocusShortcutsActive({
-        foregroundOpen: true,
-        reverseSearchOpen: false,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(false);
-  });
-
-  it('only lets Escape interrupt when no foreground input owns it', () => {
-    expect(
-      appEscapeInterruptActive({
-        inputDisabled: false,
-        reverseSearchOpen: false,
-        runPending: true,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(true);
-    expect(
-      appEscapeInterruptActive({
-        inputDisabled: false,
-        reverseSearchOpen: false,
-        runPending: false,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(false);
-    expect(
-      appEscapeInterruptActive({
-        inputDisabled: true,
-        reverseSearchOpen: false,
-        runPending: true,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(false);
-    expect(
-      appEscapeInterruptActive({
-        inputDisabled: false,
-        reverseSearchOpen: true,
-        runPending: true,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(false);
-    expect(
-      appEscapeInterruptActive({
-        inputDisabled: false,
-        reverseSearchOpen: false,
-        runPending: true,
-        slashPaletteOpen: true,
-      }),
-    ).toBe(false);
-  });
-
-  it('defers Escape interrupt whenever an Esc chord binding is visible', () => {
-    // #7496 regression: the footer advertises `[Esc s]`/`[Esc p]` whenever
-    // these controls are available, independent of the focused stream's own
-    // status — including a focused child that is still WAITING (in-flight,
-    // so its own input is NOT disabled). Deferring only for a disabled-input
-    // child left WAITING children with no chord grace window at all, so a
-    // slow `Esc s` fired an immediate interrupt and stopped the subagent.
-    expect(
-      shouldDeferEscapeInterruptForMetaChord({
-        shortcutModifierLabel: 'Esc',
-        subagentControlsAvailable: true,
-        taskControlsAvailable: false,
-      }),
-    ).toBe(true);
-    expect(
-      shouldDeferEscapeInterruptForMetaChord({
-        shortcutModifierLabel: 'Esc',
-        subagentControlsAvailable: false,
-        taskControlsAvailable: true,
-      }),
-    ).toBe(true);
-    expect(
-      shouldDeferEscapeInterruptForMetaChord({
-        shortcutModifierLabel: 'Alt',
-        subagentControlsAvailable: true,
-        taskControlsAvailable: false,
-      }),
-    ).toBe(false);
-    expect(
-      shouldDeferEscapeInterruptForMetaChord({
-        shortcutModifierLabel: 'Esc',
-        subagentControlsAvailable: false,
-        taskControlsAvailable: false,
-      }),
-    ).toBe(false);
-
-    // The exact reported repro: a focused child is WAITING (in-flight), so
-    // its own input is not disabled, yet the chord still must be deferred.
-    setParentStream(child1, root);
-    expect(
-      focusedChildInputDisabledMessage({
-        activeStreamId: child1,
-        parentStream: parentStream.get(),
-        status: STREAM_STATUS.WAITING,
-      }),
-    ).toBeUndefined();
-    expect(
-      shouldDeferEscapeInterruptForMetaChord({
-        shortcutModifierLabel: 'Esc',
-        subagentControlsAvailable: true,
-        taskControlsAvailable: false,
-      }),
-    ).toBe(true);
-  });
-
-  it('parses stripped meta shortcut digits', () => {
-    expect(digitFromMetaShortcut('1')).toBe(1);
-    expect(digitFromMetaShortcut('9')).toBe(9);
-    expect(digitFromMetaShortcut('0')).toBeUndefined();
-    expect(digitFromMetaShortcut('10')).toBeUndefined();
-    expect(digitFromMetaShortcut('p')).toBeUndefined();
-  });
-
-  it('runs Escape interrupt from the supplied current state', () => {
-    let interrupts = 0;
-    expect(
-      triggerEscapeInterrupt({
-        inputDisabled: false,
-        reverseSearchOpen: false,
-        slashPaletteOpen: false,
-        canInterruptActiveRun: () => true,
-        onInterruptActive: () => {
-          interrupts += 1;
-        },
-      }),
-    ).toBe(true);
-    expect(interrupts).toBe(1);
-
-    expect(
-      triggerEscapeInterrupt({
-        inputDisabled: false,
-        reverseSearchOpen: false,
-        slashPaletteOpen: false,
-        canInterruptActiveRun: () => false,
-        onInterruptActive: () => {
-          interrupts += 1;
-        },
-      }),
-    ).toBe(false);
-    expect(interrupts).toBe(1);
-  });
-
-  it('keeps user-opened foreground surfaces ahead of new approvals', () => {
-    expect(
-      foregroundSurfaceKind({
-        activeFormOpen: false,
-        childControlMode: 'subagents',
-        pendingApproval: true,
-        transcriptViewerOpen: false,
-      }),
-    ).toBe('childControls');
-
-    expect(
-      foregroundSurfaceKind({
-        activeFormOpen: false,
-        childControlMode: undefined,
-        pendingApproval: true,
-        transcriptViewerOpen: true,
-      }),
-    ).toBe('transcript');
-
-    expect(
-      foregroundSurfaceKind({
-        activeFormOpen: true,
-        childControlMode: undefined,
-        pendingApproval: true,
-        transcriptViewerOpen: false,
-      }),
-    ).toBe('form');
-  });
-
-  it('shows approvals when no existing foreground surface owns input', () => {
-    expect(
-      foregroundSurfaceKind({
-        activeFormOpen: false,
-        childControlMode: undefined,
-        pendingApproval: true,
-        transcriptViewerOpen: false,
-      }),
-    ).toBe('approval');
-
-    expect(
-      foregroundSurfaceKind({
-        activeFormOpen: false,
-        childControlMode: undefined,
-        pendingApproval: false,
-        transcriptViewerOpen: false,
-      }),
-    ).toBeUndefined();
-  });
-
-  it('shows stream-owned approvals only on their matching tab', () => {
-    const childApproval = {
-      payload: {
-        kind: 'bash',
-        payload: {
-          requestId: 'bash-1',
-          command: 'echo ok',
-          allowBypass: true,
-          streamId: 'child-1',
-        },
-      },
-      decide: () => undefined,
-    } satisfies PendingApproval;
-    const globalApproval = {
-      payload: {
-        kind: 'toolEdit',
-        payload: {
-          path: 'paper.tex',
-          originalContent: '',
-          proposedContent: '',
-          sourceTool: 'edit',
-        },
-      },
-      decide: () => undefined,
-    } satisfies PendingApproval;
-
-    expect(
-      approvalVisibleForActiveStream({
-        activeStreamId: 'child-1',
-        pending: childApproval,
-      }),
-    ).toBe(true);
-    expect(
-      approvalVisibleForActiveStream({
-        activeStreamId: 'root',
-        pending: childApproval,
-      }),
-    ).toBe(false);
-    expect(
-      approvalVisibleForActiveStream({
-        activeStreamId: 'root',
-        pending: globalApproval,
-      }),
-    ).toBe(true);
-  });
-
-  it('keeps hidden stream-owned approvals from taking the foreground', () => {
-    const childApproval = {
-      payload: {
-        kind: 'bash',
-        payload: {
-          requestId: 'bash-1',
-          command: 'echo ok',
-          allowBypass: true,
-          streamId: 'child-1',
-        },
-      },
-      decide: () => undefined,
-    } satisfies PendingApproval;
-
-    expect(
-      approvalVisibleForActiveStream({
-        activeStreamId: 'root',
-        pending: childApproval,
-      }),
-    ).toBe(false);
-  });
-
-  it('keeps stream focus shortcuts available for hidden approvals', () => {
-    const childApproval = {
-      payload: {
-        kind: 'bash',
-        payload: {
-          requestId: 'bash-1',
-          command: 'echo ok',
-          allowBypass: true,
-          streamId: 'child-1',
-        },
-      },
-      decide: () => undefined,
-    } satisfies PendingApproval;
-
-    expect(
-      approvalVisibleForActiveStream({
-        activeStreamId: 'root',
-        pending: childApproval,
-      }),
-    ).toBe(false);
-    expect(
-      appFocusShortcutsActive({
-        foregroundOpen: false,
-        reverseSearchOpen: false,
-        slashPaletteOpen: false,
-      }),
-    ).toBe(true);
-  });
-
-  it('labels foreground escape actions from the owning surface', () => {
-    const pending = (
-      kind: PendingApproval['payload']['kind'],
-    ): PendingApproval =>
-      ({
-        payload: { kind } as PendingApproval['payload'],
-        decide: () => undefined,
-      }) satisfies PendingApproval;
-
-    expect(
-      foregroundEscapeAction({
-        childControlEscapeAction: 'close',
-        foregroundKind: 'childControls',
-        pending: undefined,
-      }),
-    ).toBe('close');
-    expect(
-      foregroundEscapeAction({
-        childControlEscapeAction: 'back',
-        foregroundKind: 'childControls',
-        pending: undefined,
-      }),
-    ).toBe('back');
-    expect(
-      foregroundEscapeAction({ foregroundKind: 'form', pending: undefined }),
-    ).toBe('close');
-    expect(
-      foregroundEscapeAction({
-        activeFormEscapeAction: 'cancel',
-        foregroundKind: 'form',
-        pending: undefined,
-      }),
-    ).toBe('cancel');
-    expect(
-      foregroundEscapeAction({
-        foregroundKind: 'approval',
-        pending: pending('externalInquiry'),
-      }),
-    ).toBe('skip');
-    expect(
-      foregroundEscapeAction({
-        foregroundKind: 'approval',
-        pending: pending('bash'),
-      }),
-    ).toBe('cancel');
   });
 
   it('only reports a chat run interruptible after stream resolution', () => {
